requirements.txt
torch==2.2.2
torchvision==0.17.2
transformers==4.43.3
numpy==1.26.4
scipy>=1.10
Pillow
opencv-python
